# docs/api.swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Welcome to the ORY Kratos HTTP API documentation!",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
    "/auth/browser/login": {
      "get": {
        "description": "This endpoint initializes a login flow. This endpoint **should not be called from a programatic API**\nbut instead for the, for example, browser. It will redirect the user agent (e.g. browser) to the\nconfigured login UI, appending the login challenge.\n\nIf the user-agent already has a valid authentication session, the server will respond with a 302\ncode redirecting to the config value of `urls.default_return_to`.\n\nFor an in-depth look at ORY Krato's login flow, head over to: https://www.ory.sh/docs/kratos/selfservice/login",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Initialize a Login Flow",
        "operationId": "initializeLoginFlow",
        "responses": {
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
    "/profiles": {
      "get": {
        "description": "This endpoint initializes a profile update flow. This endpoint **should not be called from a programatic API**\nbut instead for the, for example, browser. It will redirect the user agent (e.g. browser) to the\nconfigured login UI, appending the login challenge.\n\nIf the user-agent does not have a valid authentication session, a 302 code will be returned which\nredirects to the initializeLoginFlow endpoint, appending this page as the return_to value.\n\nFor an in-depth look at ORY Krato's profile management flow, head over to: https://www.ory.sh/docs/kratos/selfservice/profile",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Initialize Profile Management Flow",
        "operationId": "initializeProfileManagementFlow",
        "responses": {
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      },
      "put": {
        "description": "This endpoint returns a login request's context with, for example, error details and\nother information.\n\nFor an in-depth look at ORY Krato's profile management flow, head over to: https://www.ory.sh/docs/kratos/selfservice/profile",
        "consumes": [
          "application/json",
          "application/x-www-form-urlencoded"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Complete Profile Management Flow",
        "operationId": "completeProfileManagementFlow",
        "responses": {
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
    "/profiles/requests": {
      "get": {
        "description": "This endpoint returns a profile management request's context with, for example, error details and\nother information.\n\nIt can be used from a Single Page Application or other applications running on a client device.\nThe request must be made with valid authentication cookies or it will fail!\n\nIf you wish to access this endpoint without the valid cookies (e.g. as part of a server)\nplease call this path at the admin port.\n\nFor an in-depth look at ORY Krato's profile management flow, head over to: https://www.ory.sh/docs/kratos/selfservice/profile",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Get Profile Management Request (via cookie)",
        "operationId": "getProfileManagementRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Request",
            "description": "Request should be set to the value of the `request` query parameter\nby the profile management UI.",
            "name": "request",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "profileManagementRequest",
            "schema": {
              "$ref": "#/definitions/profileManagementRequest"
            }
          },
          "302": {
            "$ref": "#/responses/emptyResponse"
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CredentialsType": {
      "description": "and so on.",
      "type": "string",
      "title": "CredentialsType  represents several different credential types, like password credentials, passwordless credentials,",
      "x-go-package": "github.com/ory/kratos/identity"
    },
...
    "FormError": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string",
          "x-go-name": "Field"
        },
        "id": {
          "$ref": "#/definitions/FormErrorCode"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
    "FormErrorCode": {
      "type": "string",
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
    "FormField": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/FormError"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "required": {
          "type": "boolean",
          "x-go-name": "Required"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        },
        "value": {
          "type": "object",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
    "FormFields": {
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/FormField"
      },
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
...
    "Identity": {
      "description": "An identity can be a real human, a service, an IoT device - everything that\ncan be described as an \"actor\" in a system.",
      "type": "object",
      "title": "Identity represents an ORY Kratos identity",
      "required": [
        "id",
        "traits"
      ],
      "properties": {
        "credentials": {
          "description": "Credentials represents all credentials that can be used for authenticating this identity.",
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/identityCredentials"
          },
          "x-go-name": "Credentials"
        },
        "id": {
          "description": "ID is a unique identifier chosen by you. It can be a URN (e.g. \"arn:aws:iam::123456789012\"),\na stringified integer (e.g. \"123456789012\"), a uuid (e.g. \"9f425a8d-7efc-4768-8f23-7647a74fdf13\"). It is up to you\nto pick a format you'd like. It is discouraged to use a personally identifiable value here, like the username\nor the email, as this field is immutable.",
          "type": "string",
          "x-go-name": "ID"
        },
        "traits": {
          "description": "Traits represent an identity's traits. The identity is able to create, modify, and delete traits\nin a self-service manner. The input will always be validated against the JSON Schema defined\nin `traits_schema_url`.",
          "type": "object",
          "x-go-name": "Traits",
          "x-go-type": {
            "import": {
              "alias": "json",
              "package": "encoding/json"
            },
            "type": "RawMessage"
          }
        },
        "traits_schema_url": {
          "description": "TraitsSchemaURL is the JSON Schema to be used for validating the identity's traits.\n\nformat: uri",
          "type": "string",
          "x-go-name": "TraitsSchemaURL"
        }
      },
      "x-go-package": "github.com/ory/kratos/identity"
    },
    "ProfileRequestForm": {
      "type": "object",
      "properties": {
        "action": {
          "description": "Action should be used as the form action URL (\u003cform action=\"{{ .Action }}\" method=\"post\"\u003e).",
          "type": "string",
          "x-go-name": "Action"
        },
        "errors": {
          "description": "Errors contains all form errors. These will be duplicates of the individual field errors.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/FormError"
          },
          "x-go-name": "Errors"
        },
        "fields": {
          "$ref": "#/definitions/FormFields"
        },
        "method": {
          "description": "Method is the form method (e.g. POST)",
          "type": "string",
          "x-go-name": "Method"
        }
      },
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
...
    "genericError": {
      "description": "Error responses are sent when an error (e.g. unauthorized, bad request, ...) occurred.",
      "type": "object",
      "title": "Error response",
      "required": [
        "error"
      ],
      "properties": {
        "error": {
          "description": "Name is the error name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "The requested resource could not be found"
        },
        "error_code": {
          "description": "Code represents the error status code (404, 403, 401, ...).",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Code",
          "example": 404
        },
        "error_debug": {
          "description": "Debug contains debug information. This is usually not available and has to be enabled.",
          "type": "string",
          "x-go-name": "Debug",
          "example": "The database adapter was unable to find the element"
        },
        "error_hint": {
          "description": "Hint contains further information on the nature of the error.",
          "type": "string",
          "x-go-name": "Hint",
          "example": "Object with RequestID 12345 does not exist"
        }
      },
      "x-go-package": "github.com/ory/kratos"
    },
...
    "identityCredentials": {
      "description": "Credentials represents a specific credential type",
      "type": "object",
      "properties": {
        "config": {
          "description": "Config contains the concrete credential payload. This might contain the bcrypt-hashed password, or the email\nfor passwordless authentication.\n\ntype: string\nformat: binary",
          "type": "object",
          "x-go-name": "Config",
          "x-go-type": {
            "import": {
              "alias": "json",
              "package": "encoding/json"
            },
            "type": "RawMessage"
          }
        },
        "id": {
          "$ref": "#/definitions/CredentialsType"
        },
        "identifiers": {
          "description": "Identifiers represents a list of unique identifiers this credential type matches.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Identifiers"
        }
      },
      "x-go-name": "Credentials",
      "x-go-package": "github.com/ory/kratos/identity"
    },
...
    "profileManagementRequest": {
      "description": "This request is used when an identity wants to update profile information\n(especially traits) in a selfservice manner.\n\nFor more information head over to: https://www.ory.sh/docs/kratos/selfservice/profile",
      "type": "object",
      "title": "ProfileManagementRequest presents a profile management request",
      "properties": {
        "expires_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ExpiresAt"
        },
        "form": {
          "$ref": "#/definitions/ProfileRequestForm"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "identity": {
          "$ref": "#/definitions/Identity"
        },
        "issued_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "IssuedAt"
        },
        "request_url": {
          "type": "string",
          "x-go-name": "RequestURL"
        }
      },
      "x-go-name": "ProfileManagementRequest",
      "x-go-package": "github.com/ory/kratos/selfservice"
    },
...
  },
  "responses": {
    "emptyResponse": {
      "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is\ntypically 201."
    }
  },
...
}